% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_helpers.R
\name{appendvar}
\alias{appendvar}
\title{Append mapped terms to a metadata variable}
\usage{
appendvar(var, val, filtv, m)
}
\arguments{
\item{var}{The variable in the metadata to assess.}

\item{val}{Character string for regex matching.}

\item{filtv}{Vector of character strings to filter on.}

\item{m}{Metadata for term lookup (data.frame).}
}
\value{
The result of assessing a regex match on a metadata variable.
}
\description{
Appends new variable data to a metadata variable, preserving current
variable terms.
}
\seealso{
get_filt, get_pstr_neg, get_pstr
}
