% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beadarray_cgctrlmetrics.R
\name{bathresh}
\alias{bathresh}
\title{Get BeadArray control outcomes from a matrix of metric signals}
\usage{
bathresh(rm, pass.lab = "PASS", fail.lab = "FAIL")
}
\arguments{
\item{rm}{BeadArray signals returned by `bactrl()` 
(matrix, rows = samples, cols = metrics).}

\item{pass.lab}{Label for passing samples (character, "PASS").}

\item{fail.lab}{Label for failing samples (character, "FAIL").}
}
\value{
Matrix of threshold assessments, either 'FAIL' or 'PASS', 
of same dimensions as input matrix.
}
\description{
Apply BeadArray thresholds to determine passing and failing samples. 
This function takes a matrix of signals and performs controls using 
minimum signal thresholds from the BeadArray Controls Reporter 
Software Guide ewastools.
}
\seealso{
bathresh
}
