% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_helpers.R
\name{get_filt}
\alias{get_filt}
\title{Get the outcome of pattern matching a string with regex patterns}
\usage{
get_filt(
  v,
  m = md,
  filtrel = "|",
  nfilt = FALSE,
  ntfilt = "",
  ptfilt = "",
  varl = c("gsm_title", "sample_type", "disease_state", "anatomic_location", "misc")
)
}
\arguments{
\item{v}{Character string or vector of such strings.}

\item{m}{Metadata used for pattern matching (data.frame).}

\item{filtrel}{Logical symbol joining each regex pattern (default "|").}

\item{nfilt}{Whether to also use negative lookup}

\item{ntfilt}{Specific character strings of terms to filter on, or for which
to use the negative lookup/regex negation (default "").}

\item{ptfilt}{Additional positive/affirmative match terms to filter (default "").}
}
\value{
The result of assessing a regex match on a metadata variable.
}
\description{
Does pattern matching for regex patterns constructed from a character
string or vector of such character strings.
}
\seealso{
appendvar
}
