% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex_helpers.R
\name{get_pstr_neg}
\alias{get_pstr_neg}
\title{Get the negation regex patterns for a given regex pattern}
\usage{
get_pstr_neg(pstr)
}
\arguments{
\item{pstr}{Regex patterns for matching, or output from `get_pstr()`.}
}
\value{
Regex for negation of indicated patterns.
}
\description{
Does progressive capitalization on values separated by spaces
for each value in v, appends flanking '.*' (matches any char)
for each value in v, appends "|" OR conditional separator
}
\seealso{
get_pstr, get_filt, appendvar
}
