% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{getblocks}
\alias{getblocks}
\title{Get index blocks, with remainder handling}
\usage{
getblocks(slength, bsize = 50)
}
\arguments{
\item{slength}{Length of the index vector (integer)}

\item{bsize}{The maximum block size (integer, default = 50).}
}
\value{
List of index blocks.
}
\description{
Get index blocks, with remainder handling. This is used by scripts
that process large data objects processively, in chunks/blocks of 
sample indices. Note the default block size of 50 is recommended for
processes using entire array's worth of data (e.g. from an RGChannel set,
getRed, getGreen, getMeth, getUnmeth, etc.), as test showed diminished
performance with greater or lesser samples on typical compute systems.
}
