% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/beadarray_cgctrlmetrics.R
\name{bactrl}
\alias{bactrl}
\title{Get control probe metrics from matrices of red/grn signal}
\usage{
bactrl(rs, gs, cdf, baseline = 3000, biotin.baseline = 1)
}
\arguments{
\item{rs}{Red signal data (data.frame, columns are probes, rows are samples, 
column names are addresses, rownames are samples/GSM IDs).}

\item{gs}{Green signal data (data.frame, columns are probes, rows are samples, 
column names are addresses, rownames are samples/GSM IDs).}

\item{cdf}{Control probe annotations (data.frame, cols = properties, rows = probes).}

\item{baseline}{Baseline measure for signals (integer, 3000).}

\item{biotin.baseline}{Baseline to use for biotin controls (integer, 1).}
}
\value{
Matrix of BeadArray signal values
}
\description{
This resource takes guidance from 3 key sources: the Illumina GenomeStudio Methylation
}
\seealso{
bathresh
}
